Add -Wconversion warning to exodus build and resolve its hundreds of warnings about implicit narrowing conversions (e.g. unsigned int to signed int) either by refactoring or adding specific static_cast<type_xxx>(yyy). Now there are no hidden conversions between signed and unsigned ints and doubles etc. Increases libexodus resiliance. Suppress exodus Ctrl+C interactive handler when there is no terminal and instead just abort/terminate

+++ exodus/libexodus/exodus/cargs.h
@@ -34,7 +34,8 @@ class Cargs {
 
 	//return argc as int
 	int argc() {
-		return argv_.size() - 1;
+		//warning: conversion from ‘std::vector<char*>::size_type’ {aka ‘long unsigned int’} to ‘int’ may change value [-Wconversion]
+		return static_cast<int>(argv_.size()) - 1;
 	}
 
 	//return argv_ as char*[]
